chore: Setup pay module, move routes
